Refactor to makeFisherySprite, dispatch when drag
