+++ client/store/fish.js
@@ -5,21 +5,6 @@ const SET_FISHES = 'SET_FISHES'
 const UPDATE_FISH = 'UPDATE_FISH'
 const REMOVE_FISH = 'REMOVE_FISH' // for client-side rendering
 
-/*export const setFishes = (fishes, resources, spritePath) => {
-   const clientFishes = fishes.map(fish => {
-   if (!fish.sprite)
-   return {
-   ...fish,
-   sprite: new Sprite(resources[`${spritePath}/fishes.png`].texture)
-   }
-   else return fish
-   })
-   return {
-   type: SET_FISHES,
-   fishes: clientFishes
-   }
-   } */
-
 export const setFishes = fishes => ({type: SET_FISHES, fishes})
 export const updateFish = fish => ({type: UPDATE_FISH, fish})
