+++ client/store/fisheries.js
@@ -1,4 +1,5 @@
 import {Sprite} from 'pixi.js'
+import {makeFisherySprite} from '../script/makeFisherySprite'
 
 const SET_FISHERIES = 'SET_FISHERIES'
 
@@ -7,6 +8,10 @@ export const setFisheries = fisheries => ({type: SET_FISHERIES, fisheries})
 export default function(state = [], action) {
   switch (action.type) {
     case SET_FISHERIES:
+      action.fisheries.forEach(fishery => {
+        if (!fishery.sprite) fishery.sprite = makeFisherySprite(fishery)
+      })
+
       return action.fisheries
     default:
       return state
